Minor refactor of finished components within homepage
