+++ src/containers/app/components/console/console.scss
@@ -1,0 +1,76 @@
+@import 'variables';
+
+$console-width: 680px;
+$console-height: 240px;
+$console-header-height: 40px;
+$console-header-button: $console-header-height / 2.5;
+
+.console-container {
+  position: relative;
+  width: 100%;
+  @include desktop-up {
+    width: $console-width;
+  }
+  .version {
+    text-align: center;
+    font-weight: $extra-bold-weight;
+    font-size: 0.8rem;
+    color: rgba($white, 0.25);
+    @include capped-text(1px);
+    margin-bottom: $medium-space;
+  }
+  .console {
+    background: $dark-gray;
+    border-radius: $extra-border-radius;
+    overflow: hidden;
+    width: 100%;
+    height: $console-height / 1.25;
+    display: block;
+    box-shadow: 0px $small-space $large-space rgba($black, 0.25);
+    @include desktop-up {
+      height: $console-height;
+    }
+    .header {
+      width: 100%;
+      height: $console-header-height;
+      background: $black;
+      overflow: hidden;
+      span {
+        float: left;
+        display: block;
+        width: $console-header-button;
+        height: $console-header-button;
+        background: rgba($white, 0.15);
+        margin-left: $small-space;
+        margin-top: ($console-header-height - $console-header-button) / 2;
+        border-radius: $round-radius;
+        &:first-of-type {
+          margin-left: $medium-space;
+        }
+      }
+    }
+    .content {
+      position: relative;
+      padding: $medium-space;
+      p {
+        color: $white;
+        letter-spacing: 0.5px;
+        margin: 0px;
+        word-wrap: break-word;
+        // NOTE: Bold weight (700) is not available on GWF... we're faking it below, maybe find a replacement?
+        font-family: $monospace-font;
+        font-weight: 700;
+        .screenname {
+          user-select: none;
+          color: $success-color;
+        }
+      }
+    }
+    button {
+      position: absolute;
+      bottom: 0px;
+      right: 0px;
+      margin: $medium-space;
+    }
+  }
+}
