+++ src/containers/app/routes/static/homepage/hero/hero.scss
@@ -1,0 +1,65 @@
+@import 'variables';
+
+$hero-height: 800px;
+
+#hero {
+  position: relative;
+  height: auto;
+  @include desktop-up {
+    height: $hero-height;
+  }
+  &:after {
+    position: absolute;
+    bottom: 0px;
+    left: 0px;
+    width: 100%;
+    height: 10px;
+    background: rgba($white, 0.25);
+    z-index: -1;
+    content: '';
+  }
+  #logo {
+    display: block;
+    width: 160px;
+    margin: $large-space auto;
+    @include desktop-up {
+      margin: $xlarge-space auto;
+      width: 200px;
+    }
+  }
+  #tagline {
+    text-align: center;
+    color: $white;
+    font-size: 1.35rem;
+    @include desktop-up {
+      font-size: 2.25rem;
+    }
+  }
+  #description {
+    text-align: center;
+    font-size: 1.1rem;
+    line-height: 2rem;
+    color: $white;
+    width: 100%;
+    display: block;
+    margin: $medium-space auto 0px auto;
+    @include tablet-up {
+      width: 80%;
+    }
+    @include desktop-up {
+      line-height: 2.5rem;
+      font-size: 1.25rem;
+    }
+    @include widescreen-up {
+      width: 55%;
+    }
+  }
+  #steps {
+    margin-top: $large-space;
+    padding-bottom: $medium-space;
+    @include desktop-up() {
+      margin-top: $xlarge-space;
+      padding-bottom: 0px;
+    }
+  }
+}
