+++ src/containers/app/routes/static/homepage/homepage.scss
@@ -1,10 +1,6 @@
 @import 'variables';
 
 #homepage {
-  @import 'hero';
-  @import 'testimonials';
-  @import 'mission';
-  @import 'pillars';
   @import 'process';
   @import 'timeline';
   @import 'footer';
