+++ src/containers/app/routes/static/homepage/mission/mission.scss
@@ -1,3 +1,5 @@
+@import 'variables';
+
 #mission {
   padding-top: $large-space;
   @include desktop-up {
