+++ src/containers/app/routes/static/homepage/pillars/pillars.scss
@@ -1,3 +1,5 @@
+@import 'variables';
+
 #pillars {
   padding: $medium-space 0px;
   @include desktop-up {
