+++ src/containers/app/routes/static/homepage/testimonials/index.js
@@ -1,7 +1,15 @@
 import React from 'react';
 import { Row, Column, Container } from 'openmined-ui';
 
-import ExternalLink from '../../../components/external-link';
+import ExternalLink from '../../../../components/external-link';
+
+import './testimonials.css';
+
+const Testimonial = ({ name, link, image }) => (
+  <ExternalLink to={link} className="testimonial">
+    <img src={image} className="image" alt={name} />
+  </ExternalLink>
+);
 
 const Testimonials = ({ testimonials }) => (
   <div id="testimonials">
@@ -9,10 +17,8 @@ const Testimonials = ({ testimonials }) => (
       <Row>
         <Column sizes={{ small: 12, xlarge: 10 }} offsets={{ xlarge: 1 }}>
           <div className="testimonial-list">
-            {testimonials.map(({ name, link, image }) => (
-              <ExternalLink to={link} className="testimonial" key={name}>
-                <img src={image} className="image" alt={name} />
-              </ExternalLink>
+            {testimonials.map((testimonial, index) => (
+              <Testimonial {...testimonial} key={index} />
             ))}
           </div>
         </Column>
