+++ src/containers/app/routes/static/homepage/testimonials/testimonials.scss
@@ -1,3 +1,5 @@
+@import 'variables';
+
 #testimonials {
   display: none;
   @include desktop-up {
